close #635 close #629 Refactor mega models file of laws to different files, allow reverting to previous behavior if needed by waffle switch
