+++ agendas/models.py
@@ -8,21 +8,22 @@ from django.db import connection
 from django.db import models
 from django.db.models import Sum, Q, Count, F
 from django.utils.translation import ugettext_lazy as _
-from django.contrib.contenttypes.models import ContentType
+
 from django.core.cache import cache
 from django.core.urlresolvers import reverse
 from django.conf import settings
 
 from django.contrib.auth.models import User
 from actstream.models import Follow
-from laws.models import VoteAction, Vote
+from laws.models.vote_action import VoteAction
+from laws.models.vote import Vote
 from mks.models import Party, Member, Knesset, Membership
 import queries
 
 from tagging.models import Tag
 
 AGENDAVOTE_SCORE_CHOICES = (
-    ('',_("Not selected")),
+    ('', _("Not selected")),
     (-1.0, _("Opposes fully")),
     (-0.5, _("Opposes partially")),
     (0.0, _("Agnostic")),
@@ -30,12 +31,13 @@ AGENDAVOTE_SCORE_CHOICES = (
     (1.0, _("Complies fully")),
 )
 IMPORTANCE_CHOICES = (
-    ('',_("Not selected")),
+    ('', _("Not selected")),
     (0.0, _("Marginal Importance")),
     (0.3, _("Medium Importance")),
     (0.6, _("High Importance")),
     (1.0, _("Very High Importance")),
 )
+
 
 class UserSuggestedVote(models.Model):
     agenda = models.ForeignKey('Agenda', related_name='user_suggested_votes')
@@ -45,12 +47,13 @@ class UserSuggestedVote(models.Model):
     sent_to_editor = models.BooleanField(default=False)
 
     class Meta:
-        unique_together = ('agenda','vote','user')
+        unique_together = ('agenda', 'vote', 'user')
+
 
 class AgendaVoteManager(models.Manager):
     db_month_trunc_functions = {
-        'sqlite3':{'monthfunc':"strftime('%%Y-%%m-01'",'nowfunc':'date()'},
-        'postgresql_psycopg2':{'monthfunc':"date_trunc('month'",'nowfunc':'now()'}
+        'sqlite3': {'monthfunc': "strftime('%%Y-%%m-01'", 'nowfunc': 'date()'},
+        'postgresql_psycopg2': {'monthfunc': "date_trunc('month'", 'nowfunc': 'now()'}
     }
 
     def compute_all(self):
@@ -62,7 +65,7 @@ class AgendaVoteManager(models.Manager):
 
         mk_query = queries.BASE_MK_QUERY % db_functions
         cursor.execute(mk_query)
-        
+
         party_query = queries.BASE_PARTY_QUERY % db_functions
         cursor.execute(party_query)
 
@@ -72,7 +75,7 @@ class AgendaVote(models.Model):
     vote = models.ForeignKey('laws.Vote', related_name='agendavotes')
     score = models.FloatField(default=0.0, choices=AGENDAVOTE_SCORE_CHOICES)
     importance = models.FloatField(default=1.0, choices=IMPORTANCE_CHOICES)
-    reasoning = models.TextField(null=True,blank=True)
+    reasoning = models.TextField(null=True, blank=True)
 
     objects = AgendaVoteManager()
 
@@ -81,38 +84,39 @@ class AgendaVote(models.Model):
 
     def get_score_header(self):
         return _('Position')
+
     def get_importance_header(self):
         return _('Importance')
 
     class Meta:
-        unique_together= ('agenda', 'vote')
+        unique_together = ('agenda', 'vote')
 
     def __unicode__(self):
-        return u"%s %s" % (self.agenda,self.vote)
+        return u"%s %s" % (self.agenda, self.vote)
 
     def update_monthly_counters(self):
-        agendaScore     = float(self.score) * float(self.importance)
-        objMonth        = dateMonthTruncate(self.vote.time)
+        agendaScore = float(self.score) * float(self.importance)
+        objMonth = dateMonthTruncate(self.vote.time)
 
-        summaryObjects  = list(SummaryAgenda.objects.filter( agenda=self.agenda,
-                                                            month=objMonth).all())
+        summaryObjects = list(SummaryAgenda.objects.filter(agenda=self.agenda,
+                                                           month=objMonth).all())
 
-        agendaSummary   = None
-        if not filter(lambda summary:summary.summary_type=='AG',summaryObjects):
-            agendaSummary   = SummaryAgenda(month=objMonth,
-                                            agenda=self.agenda,
-                                            summary_type='AG',
-                                            score=abs(agendaScore),
-                                            votes=1)
+        agendaSummary = None
+        if not filter(lambda summary: summary.summary_type == 'AG', summaryObjects):
+            agendaSummary = SummaryAgenda(month=objMonth,
+                                          agenda=self.agenda,
+                                          summary_type='AG',
+                                          score=abs(agendaScore),
+                                          votes=1)
         else:
-            existingAgenda = filter(lambda summary:summary.summary_type=='AG',summaryObjects)[0]
+            existingAgenda = filter(lambda summary: summary.summary_type == 'AG', summaryObjects)[0]
             existingAgenda.votes += 1
             existingAgenda.score += abs(agendaScore)
             existingAgenda.save()
 
-        agendasByMk     = dict(map(lambda summary:(summary.mk_id,summary),
-                                   filter(lambda summary:summary.summary_type=='MK',
-                                          summaryObjects)))
+        agendasByMk = dict(map(lambda summary: (summary.mk_id, summary),
+                               filter(lambda summary: summary.summary_type == 'MK',
+                                      summaryObjects)))
         newObjects = []
         if agendaSummary:
             newObjects.append(agendaSummary)
@@ -132,14 +136,17 @@ class AgendaVote(models.Model):
             else:
                 voters[vote_action.type].append(vote_action.member_id)
 
-        SummaryAgenda.objects.filter(agenda=self.agenda,mk_id__in=voters['for'],month=objMonth).update(votes=F('votes') + 1, score=F('score')+agendaScore, for_votes=F('for_votes') + 1)
-        SummaryAgenda.objects.filter(agenda=self.agenda,mk_id__in=voters['against'],month=objMonth).update(votes=F('votes') + 1, score=F('score')-agendaScore, against_votes=F('against_votes') + 1)
+        SummaryAgenda.objects.filter(agenda=self.agenda, mk_id__in=voters['for'], month=objMonth).update(
+            votes=F('votes') + 1, score=F('score') + agendaScore, for_votes=F('for_votes') + 1)
+        SummaryAgenda.objects.filter(agenda=self.agenda, mk_id__in=voters['against'], month=objMonth).update(
+            votes=F('votes') + 1, score=F('score') - agendaScore, against_votes=F('against_votes') + 1)
         if newObjects:
             SummaryAgenda.objects.bulk_create(newObjects)
 
-    def save(self,*args,**kwargs):
-        super(AgendaVote,self).save(*args,**kwargs)
+    def save(self, *args, **kwargs):
+        super(AgendaVote, self).save(*args, **kwargs)
         self.update_monthly_counters()
+
 
 class AgendaMeeting(models.Model):
     agenda = models.ForeignKey('Agenda', related_name='agendameetings')
@@ -153,6 +160,7 @@ class AgendaMeeting(models.Model):
 
     def get_score_header(self):
         return _('Importance')
+
     def get_importance_header(self):
         return ''
 
@@ -160,7 +168,8 @@ class AgendaMeeting(models.Model):
         unique_together = ('agenda', 'meeting')
 
     def __unicode__(self):
-        return u"%s %s" % (self.agenda,self.meeting)
+        return u"%s %s" % (self.agenda, self.meeting)
+
 
 class AgendaBill(models.Model):
     agenda = models.ForeignKey('Agenda', related_name='agendabills')
@@ -174,6 +183,7 @@ class AgendaBill(models.Model):
 
     def get_score_header(self):
         return _('Position')
+
     def get_importance_header(self):
         return _('Importance')
 
@@ -181,7 +191,8 @@ class AgendaBill(models.Model):
         unique_together = ('agenda', 'bill')
 
     def __unicode__(self):
-        return u"%s %s" % (self.agenda,self.bill)
+        return u"%s %s" % (self.agenda, self.bill)
+
 
 def get_top_bottom(lst, top, bottom):
     """
@@ -189,13 +200,13 @@ def get_top_bottom(lst, top, bottom):
     Edge conditions are handled gracefuly.
     Input list should be ascending so that top is at the end.
     """
-    if len(lst) < top+bottom:
-        delta = top+bottom - len(lst)
-        bottom = bottom - int(math.floor(delta/2))
-        if delta%2:
-            top = top - int(math.floor(delta/2)) -1
+    if len(lst) < top + bottom:
+        delta = top + bottom - len(lst)
+        bottom = bottom - int(math.floor(delta / 2))
+        if delta % 2:
+            top = top - int(math.floor(delta / 2)) - 1
         else:
-            top = top - int(math.floor(delta/2))
+            top = top - int(math.floor(delta / 2))
     if top and bottom:
         top_lst = lst[-top:]
         bottom_lst = lst[:bottom]
@@ -209,12 +220,11 @@ def get_top_bottom(lst, top, bottom):
         top_lst = []
         bottom_lst = []
 
-    return {'top':top_lst,
-            'bottom':bottom_lst}
+    return {'top': top_lst,
+            'bottom': bottom_lst}
 
 
 class AgendaManager(models.Manager):
-
     def get_selected_for_instance(self, instance, user=None, top=3, bottom=3):
         # Returns interesting agendas for model instances such as: member, party
         agendas = list(self.get_relevant_for_user(user))
@@ -228,33 +238,33 @@ class AgendaManager(models.Manager):
         return agendas
 
     def get_relevant_for_mk(self, mk, agendaId):
-        agendas = AgendaVote.objects.filter(agenda__id=agendaId,vote__votes__id=mk).distinct()
+        agendas = AgendaVote.objects.filter(agenda__id=agendaId, vote__votes__id=mk).distinct()
         return agendas
 
     def get_relevant_for_user(self, user):
         if user == None or not user.is_authenticated():
-            agendas = Agenda.objects.filter(is_public=True)\
-                                    .order_by('-num_followers')\
-                                    .prefetch_related('agendavotes')
+            agendas = Agenda.objects.filter(is_public=True) \
+                .order_by('-num_followers') \
+                .prefetch_related('agendavotes')
         elif user.is_superuser:
-            agendas = Agenda.objects.all().order_by('-num_followers')\
-                                          .prefetch_related('agendavotes')
+            agendas = Agenda.objects.all().order_by('-num_followers') \
+                .prefetch_related('agendavotes')
         else:
             agendas = Agenda.objects.filter(Q(is_public=True) |
-                                            Q(editors=user))\
-                                    .order_by('-num_followers')\
-                                    .prefetch_related('agendavotes')\
-                                    .distinct()
+                                            Q(editors=user)) \
+                .order_by('-num_followers') \
+                .prefetch_related('agendavotes') \
+                .distinct()
         return agendas
 
     def get_possible_to_suggest(self, user, vote):
         if user == None or not user.is_authenticated():
             agendas = False
         else:
-            agendas = Agenda.objects.filter(is_public=True)\
-                            .exclude(editors=user)\
-                            .exclude(agendavotes__vote=vote)\
-                            .distinct()
+            agendas = Agenda.objects.filter(is_public=True) \
+                .exclude(editors=user) \
+                .exclude(agendavotes__vote=vote) \
+                .distinct()
         return agendas
 
     def get_mks_values(self):
@@ -265,20 +275,20 @@ class AgendaManager(models.Manager):
             newAgendaMkVotes = {}
             # generates a set of all the current mk ids that have ever voted for any agenda
             # its not perfect, but its better than creating another query to generate all known mkids
-            allMkIds = set(map(itemgetter(0),chain.from_iterable(q.values())))
-            for agendaId,agendaVotes in q.items():
+            allMkIds = set(map(itemgetter(0), chain.from_iterable(q.values())))
+            for agendaId, agendaVotes in q.items():
                 # the newdict will have 0's for each mkid, the update will change the value for known mks
-                newDict = {}.fromkeys(allMkIds,(0,0,0))
-                newDict.update(dict(map(lambda (mkid,score,volume,numvotes):(mkid,(score,volume,numvotes)),agendaVotes)))
-                newAgendaMkVotes[agendaId]=newDict.items()
+                newDict = {}.fromkeys(allMkIds, (0, 0, 0))
+                newDict.update(
+                    dict(map(lambda (mkid, score, volume, numvotes): (mkid, (score, volume, numvotes)), agendaVotes)))
+                newAgendaMkVotes[agendaId] = newDict.items()
             mks_values = {}
             for agenda_id, scores in newAgendaMkVotes.items():
                 mks_values[agenda_id] = \
                     map(lambda x: (x[1][0], dict(score=x[1][1][0], rank=x[0], volume=x[1][1][1], numvotes=x[1][1][2])),
-                        enumerate(sorted(scores,key=lambda x:x[1][0],reverse=True), 1))
+                        enumerate(sorted(scores, key=lambda x: x[1][0], reverse=True), 1))
             cache.set('agendas_mks_values', mks_values, 1800)
         return mks_values
-
 
     # def get_mks_values(self,ranges=None):
     #     if ranges is None:
@@ -319,11 +329,12 @@ class AgendaManager(models.Manager):
     def get_all_party_values(self):
         return queries.getAllAgendaPartyVotes()
 
+
 class Agenda(models.Model):
     name = models.CharField(max_length=200)
-    description = models.TextField(null=True,blank=True)
+    description = models.TextField(null=True, blank=True)
     editors = models.ManyToManyField('auth.User', related_name='agendas')
-    votes = models.ManyToManyField('laws.Vote',through=AgendaVote)
+    votes = models.ManyToManyField('laws.Vote', through=AgendaVote)
     public_owner_name = models.CharField(max_length=100)
     is_public = models.BooleanField(default=False)
     num_followers = models.IntegerField(default=0)
@@ -341,7 +352,7 @@ class Agenda(models.Model):
         unique_together = (("name", "public_owner_name"),)
 
     def __unicode__(self):
-        return u"%s %s %s" % (self.name,_('edited by'),self.public_owner_name)
+        return u"%s %s %s" % (self.name, _('edited by'), self.public_owner_name)
 
     @models.permalink
     def get_absolute_url(self):
@@ -354,7 +365,7 @@ class Agenda(models.Model):
     def member_score(self, member):
         values = self.get_mks_values(mks=[member])
         if values:
-            if len(values)>1:
+            if len(values) > 1:
                 raise Member.MultipleObjectsReturned
             return values[0][1]['score']
         else:
@@ -366,8 +377,8 @@ class Agenda(models.Model):
         qs = AgendaVote.objects.filter(
             agenda=self, vote__actions__member__in=party.members.all(),
             vote__actions__type__in=['against', 'for']).extra(
-                select={'weighted_score': 'agendas_agendavote.score*agendas_agendavote.importance'}
-            ).values_list('weighted_score', 'vote__actions__type')
+            select={'weighted_score': 'agendas_agendavote.score*agendas_agendavote.importance'}
+        ).values_list('weighted_score', 'vote__actions__type')
 
         for_score = 0
         against_score = 0
@@ -378,7 +389,7 @@ class Agenda(models.Model):
             else:
                 for_score += score
 
-        #max_score = sum([abs(x) for x in self.agendavotes.values_list('score', flat=True)]) * party.members.count()
+        # max_score = sum([abs(x) for x in self.agendavotes.values_list('score', flat=True)]) * party.members.count()
         # To save the queries, make sure to pass prefetch/select related
         # Removed the values call, so that we can utilize the prefetched stuf
         # This reduces the number of queries when called for example from
@@ -397,8 +408,8 @@ class Agenda(models.Model):
         qs = AgendaVote.objects.filter(
             agenda=self, vote__actions__member__in=candidate_list.member_ids,
             vote__actions__type__in=['against', 'for']).extra(
-                select={'weighted_score': 'agendas_agendavote.score*agendas_agendavote.importance'}
-            ).values_list('weighted_score', 'vote__actions__type')
+            select={'weighted_score': 'agendas_agendavote.score*agendas_agendavote.importance'}
+        ).values_list('weighted_score', 'vote__actions__type')
 
         for_score = 0
         against_score = 0
@@ -409,7 +420,7 @@ class Agenda(models.Model):
             else:
                 for_score += score
 
-        #max_score = sum([abs(x) for x in self.agendavotes.values_list('score', flat=True)]) * party.members.count()
+        # max_score = sum([abs(x) for x in self.agendavotes.values_list('score', flat=True)]) * party.members.count()
         # To save the queries, make sure to pass prefetch/select related
         # Removed the values call, so that we can utilize the prefetched stuf
         # This reduces the number of queries when called for example from
@@ -422,24 +433,24 @@ class Agenda(models.Model):
         else:
             return 0.0
 
-    def related_mk_votes(self,member):
+    def related_mk_votes(self, member):
         # Find all votes that
         #   1) This agenda is ascribed to
         #   2) the member participated in and either voted for or against
         # for_votes      = AgendaVote.objects.filter(agenda=self,vote__voteaction__member=member,vote__voteaction__type="for").distinct()
-        #against_votes   = AgendaVote.objects.filter(agenda=self,vote__voteaction__member=member,vote__voteaction__type="against").distinct()
-        vote_actions = VoteAction.objects.filter(member=member,vote__agendavotes__agenda=self)
-        all_votes = AgendaVote.objects.filter(agenda=self,vote__actions__member=member).distinct()
+        # against_votes   = AgendaVote.objects.filter(agenda=self,vote__voteaction__member=member,vote__voteaction__type="against").distinct()
+        vote_actions = VoteAction.objects.filter(member=member, vote__agendavotes__agenda=self)
+        all_votes = AgendaVote.objects.filter(agenda=self, vote__actions__member=member).distinct()
         # TODO: improve ugly code below
         member_votes = list()
         for member_vote in all_votes:
             for vote_action in vote_actions:
                 if (vote_action.vote == member_vote.vote):
-                    member_votes.insert(0,member_vote)
+                    member_votes.insert(0, member_vote)
                     member_votes[0].voteaction = vote_action
 
         return member_votes
-        #return AgendaVote.objects.filter(agenda=self,vote__voteaction__member=mk).distinct()
+        # return AgendaVote.objects.filter(agenda=self,vote__voteaction__member=mk).distinct()
 
     def selected_instances(self, cls, top=3, bottom=3):
         instances = list(cls.objects.all())
@@ -461,9 +472,9 @@ class Agenda(models.Model):
                 return None  # might as well not filter at all
             query_fields = {}
             if r[0]:
-                query_fields[start_fieldname + '__gte']=r[0]
+                query_fields[start_fieldname + '__gte'] = r[0]
             if r[1]:
-                query_fields[end_fieldname + '__lt']=r[1]
+                query_fields[end_fieldname + '__lt'] = r[1]
             filter_list.append(Q(**query_fields))
 
         if len(filter_list) == 1:
@@ -487,13 +498,13 @@ class Agenda(models.Model):
     def get_mks_values(self, ranges=None, mks=None):
         if ranges is None:
             only_current_mks = True
-            ranges = [[dateMonthTruncate(Knesset.objects.current_knesset().start_date),None]]
+            ranges = [[dateMonthTruncate(Knesset.objects.current_knesset().start_date), None]]
         else:
             only_current_mks = False
         mks_values = False
         mk_ids = [mk.id for mk in mks] if mks else []
 
-        fullRange = ranges == [[None,None]]
+        fullRange = ranges == [[None, None]]
         if fullRange:
             mks_values = cache.get('agenda_%d_mks_values' % self.id)
             if mks_values and mks:
@@ -518,46 +529,47 @@ class Agenda(models.Model):
             for r in ranges:
                 summariesForRange = defaultdict(list)
                 for s in summaries:
-                    if (not r[0] or s.month>=r[0]) and \
-                        (not r[1] or s.month<r[1]):
+                    if (not r[0] or s.month >= r[0]) and \
+                            (not r[1] or s.month < r[1]):
                         summariesForRange[s.summary_type].append(s)
                 summariesForRanges.append(summariesForRange)
 
             # compute agenda measures, store results per MK
-            mk_results = dict(map(lambda mk_id:(mk_id,[]),mk_ids))
+            mk_results = dict(map(lambda mk_id: (mk_id, []), mk_ids))
             for summaries in summariesForRanges:
-                agenda_data             = summaries['AG']
-                total_votes             = sum(map(attrgetter('votes'),agenda_data))
-                total_for_votes         = sum(map(attrgetter('for_votes'),agenda_data))
-                total_against_votes     = sum(map(attrgetter('against_votes'),agenda_data))                
-                total_score             = sum(map(attrgetter('score'),agenda_data))
-                current_mks_data        = indexby(summaries['MK'],attrgetter('mk_id'))
+                agenda_data = summaries['AG']
+                total_votes = sum(map(attrgetter('votes'), agenda_data))
+                total_for_votes = sum(map(attrgetter('for_votes'), agenda_data))
+                total_against_votes = sum(map(attrgetter('against_votes'), agenda_data))
+                total_score = sum(map(attrgetter('score'), agenda_data))
+                current_mks_data = indexby(summaries['MK'], attrgetter('mk_id'))
                 # calculate results per mk
-                rangeMkResults          = []
+                rangeMkResults = []
                 for mk_id in mk_results.keys():
-                    mk_data     = current_mks_data[mk_id]
+                    mk_data = current_mks_data[mk_id]
                     if mk_data:
-                        mk_votes            = sum(map(attrgetter('votes'),mk_data))
-                        mk_for_votes        = sum(map(attrgetter('for_votes'),mk_data))
-                        mk_against_votes    = sum(map(attrgetter('against_votes'),mk_data))
-                        mk_volume           = 100*mk_votes/total_votes
-                        mk_score            = 100*sum(map(attrgetter('score'),mk_data))/total_score if total_score != 0 else 0
-                        rangeMkResults.append((mk_id,mk_votes,mk_for_votes,mk_against_votes,mk_score,mk_volume))
+                        mk_votes = sum(map(attrgetter('votes'), mk_data))
+                        mk_for_votes = sum(map(attrgetter('for_votes'), mk_data))
+                        mk_against_votes = sum(map(attrgetter('against_votes'), mk_data))
+                        mk_volume = 100 * mk_votes / total_votes
+                        mk_score = 100 * sum(map(attrgetter('score'), mk_data)) / total_score if total_score != 0 else 0
+                        rangeMkResults.append((mk_id, mk_votes, mk_for_votes, mk_against_votes, mk_score, mk_volume))
                     else:
-                        rangeMkResults.append(tuple([mk_id]+[0]*5))
+                        rangeMkResults.append(tuple([mk_id] + [0] * 5))
                 # sort results by score descending
-                for rank,(mk_id,mk_votes,mk_for_votes,mk_against_votes,mk_score,mk_volume) in enumerate(sorted(rangeMkResults,key=itemgetter(4,0),reverse=True)):
-                    mk_range_data = dict(score=mk_score,rank=rank,volume=mk_volume,numvotes=mk_votes,numforvotes=mk_for_votes,numagainstvotes=mk_against_votes)
-                    if len(ranges)==1:
-                        mk_results[mk_id]=mk_range_data
+                for rank, (mk_id, mk_votes, mk_for_votes, mk_against_votes, mk_score, mk_volume) in enumerate(
+                        sorted(rangeMkResults, key=itemgetter(4, 0), reverse=True)):
+                    mk_range_data = dict(score=mk_score, rank=rank, volume=mk_volume, numvotes=mk_votes,
+                                         numforvotes=mk_for_votes, numagainstvotes=mk_against_votes)
+                    if len(ranges) == 1:
+                        mk_results[mk_id] = mk_range_data
                     else:
                         mk_results[mk_id].append(mk_range_data)
             if fullRange:
                 cache.set('agenda_%d_mks_values' % self.id, mks_values, 1800)
-        if len(ranges)==1:
-            mk_results = sorted(mk_results.items(),key=lambda (k,v):v['rank'])
+        if len(ranges) == 1:
+            mk_results = sorted(mk_results.items(), key=lambda (k, v): v['rank'])
         return mk_results
-
 
     def get_mks_values_old(self, knesset_number=None):
         """Return mks values.
@@ -581,7 +593,7 @@ class Agenda(models.Model):
 
     def get_party_values(self):
         party_grades = Agenda.objects.get_all_party_values()
-        return party_grades.get(self.id,[])
+        return party_grades.get(self.id, [])
 
     def get_all_party_values(self):
         return Agenda.objects.get_all_party_values()
@@ -602,47 +614,54 @@ class Agenda(models.Model):
         WHERE avti.tag_id = ati.tag_id
         """
         agenda_type_id = ContentType.objects.get_for_model(self).id
-        votes = votes.extra(select=dict(score = tag_importance_subquery),
-                            select_params = [agenda_type_id]*2)
+        votes = votes.extra(select=dict(score=tag_importance_subquery),
+                            select_params=[agenda_type_id] * 2)
         return votes.order_by('-score')[:num]
 
     def get_suggested_votes_by_controversy(self, num):
         votes = Vote.objects.filter(~Q(agendavotes__agenda=self))
-        votes = votes.extra(select=dict(score = 'controversy'))
+        votes = votes.extra(select=dict(score='controversy'))
         return votes.order_by('-score')[:num]
 
+
 SUMMARY_TYPES = (
-    ('AG','Agenda Votes'),
-    ('MK','MK Counter')
+    ('AG', 'Agenda Votes'),
+    ('MK', 'MK Counter')
 )
 
+
 class SummaryAgenda(models.Model):
-    agenda          = models.ForeignKey(Agenda, related_name='score_summaries')
-    month           = models.DateTimeField(db_index=True)
-    summary_type    = models.CharField(max_length=2, choices=SUMMARY_TYPES)
-    score           = models.FloatField(default=0.0)
-    votes           = models.BigIntegerField(default=0)
-    for_votes       = models.BigIntegerField(default=0)
-    against_votes   = models.BigIntegerField(default=0)
-    mk              = models.ForeignKey(Member,blank=True, null=True, related_name='agenda_summaries')
-    db_created      = models.DateTimeField(auto_now_add=True)
-    db_updated      = models.DateTimeField(auto_now=True)
+    agenda = models.ForeignKey(Agenda, related_name='score_summaries')
+    month = models.DateTimeField(db_index=True)
+    summary_type = models.CharField(max_length=2, choices=SUMMARY_TYPES)
+    score = models.FloatField(default=0.0)
+    votes = models.BigIntegerField(default=0)
+    for_votes = models.BigIntegerField(default=0)
+    against_votes = models.BigIntegerField(default=0)
+    mk = models.ForeignKey(Member, blank=True, null=True, related_name='agenda_summaries')
+    db_created = models.DateTimeField(auto_now_add=True)
+    db_updated = models.DateTimeField(auto_now=True)
 
     def __unicode__(self):
-        return "%s %s %s %s (%f,%d)" % (str(self.agenda_id),str(self.month),self.summary_type,str(self.mk_id) if self.mk else u'n/a',self.score,self.votes)
+        return "%s %s %s %s (%f,%d)" % (
+        str(self.agenda_id), str(self.month), self.summary_type, str(self.mk_id) if self.mk else u'n/a', self.score,
+        self.votes)
+
 
 from listeners import *
+
 
 def dateMonthTruncate(dt):
     dt = dt.replace(day=1)
     if type(dt) == datetime.datetime:
-        dt = dt.replace(hour=0,minute=0,second=0,microsecond=0)
+        dt = dt.replace(hour=0, minute=0, second=0, microsecond=0)
     else:
-        dt = datetime.datetime(year=dt.year,month=dt.month,day=1)
+        dt = datetime.datetime(year=dt.year, month=dt.month, day=1)
     return dt
 
-def indexby(data,fieldFunc):
+
+def indexby(data, fieldFunc):
     d = defaultdict(list)
-    for k,v in map(lambda d:(fieldFunc(d),d),data):
+    for k, v in map(lambda d: (fieldFunc(d), d), data):
         d[k].append(v)
     return d
