+++ laws/constants.py
@@ -1,4 +1,6 @@
+# encoding: utf-8
 import datetime
 
 STANDS_FOR_THRESHOLD = 0.66
 FIRST_KNESSET_START = datetime.date(1948, 5, 13)
+CONVERT_TO_DISCUSSION_HEADERS = ('להעביר את הנושא'.decode('utf8'), 'העברת הנושא'.decode('utf8'))
