+++ laws/enums.py
@@ -1,3 +1,5 @@
+from django.utils.translation import ugettext_lazy as _
+
 from knesset.enums import Enum
 
 
@@ -14,3 +16,11 @@ class BillStages(Enum):
     COMMITTEE_CORRECTIONS = u'5'
     APPROVED = u'6'
     FAILED_APPROVAL = u'-6'
+
+
+VOTE_ACTION_TYPE_CHOICES = (
+    (u'for', _('For')),
+    (u'against', _('Against')),
+    (u'abstain', _('Abstain')),
+    (u'no-vote', _('No Vote')),
+)
