+++ laws/management/commands/freeze_bills.py
@@ -1,3 +1,4 @@
+# encoding: utf-8
 from __future__ import print_function
 
 from django.core.management.base import BaseCommand
