+++ laws/templatetags/bills_tags.py
@@ -2,7 +2,6 @@
 import json
 
 from django import template
-from django.conf import settings
 
 register = template.Library()
