+++ notify/management/commands/notify.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import
 from django.core.management.base import NoArgsCommand
-from django.contrib.auth.models import User,Group
+from django.contrib.auth.models import User, Group
 from django.contrib.contenttypes.models import ContentType
 from django.contrib.sites.models import Site
 from django.utils.translation import ugettext as _
@@ -12,6 +12,7 @@ from django.core.cache import cache
 import datetime
 from optparse import make_option
 import logging
+
 logger = logging.getLogger("open-knesset.notify")
 
 from actstream.models import Follow, Action
@@ -42,9 +43,9 @@ class Command(NoArgsCommand):
 
     option_list = NoArgsCommand.option_list + (
         make_option('--daily', action='store_true', dest='daily',
-            help="send notifications to users that requested a daily update"),
+                    help="send notifications to users that requested a daily update"),
         make_option('--weekly', action='store_true', dest='weekly',
-            help="send notifications to users that requested a weekly update"))
+                    help="send notifications to users that requested a weekly update"))
 
     def agenda_update(self, agenda):
         ''' generate the general update email for this agenda.
@@ -54,12 +55,12 @@ class Command(NoArgsCommand):
         mks = agenda.selected_instances(Member)
         template_name = 'notify/agenda_update'
         update_txt = render_to_string(template_name + '.txt',
-                                      {'mks':mks,
-                                       'domain':self.domain})
+                                      {'mks': mks,
+                                       'domain': self.domain})
         update_html = render_to_string(template_name + '.html',
-                                       {'mks':mks,
-                                        'domain':self.domain})
-        return (update_txt,update_html)
+                                       {'mks': mks,
+                                        'domain': self.domain})
+        return (update_txt, update_html)
 
     @classmethod
     def get_model_headers(cls, model):
@@ -76,9 +77,9 @@ class Command(NoArgsCommand):
 
     def get_email_for_user(self, user):
         ''' return the body text and html for a user's email '''
-        updates = dict(zip(self.update_models, ([] for x in self.update_models))) # will contain the updates to be sent
+        updates = dict(zip(self.update_models, ([] for x in self.update_models)))  # will contain the updates to be sent
         updates_html = dict(zip(self.update_models, ([] for x in self.update_models)))
-        follows = Follow.objects.filter(user=user) # everything this user is following
+        follows = Follow.objects.filter(user=user)  # everything this user is following
         # sometime a user follows something several times. we want to filter that out:
         follows = set([f.actor for f in follows])
         for f in follows:
@@ -96,58 +97,67 @@ class Command(NoArgsCommand):
             if model_class in updates:
                 key = model_class
             else:
-                key = None # put all updates for 'other' classes at the 'None' group
-            try: # get actions that happened since last update
+                key = None  # put all updates for 'other' classes at the 'None' group
+            try:  # get actions that happened since last update
                 last_sent = LastSent.objects.get(user=user, content_type=content_type, object_pk=f.id)
                 last_sent_time = last_sent.time
-                stream = Action.objects.filter(actor_content_type = content_type,
-                                               actor_object_id = f.id,
+                stream = Action.objects.filter(actor_content_type=content_type,
+                                               actor_object_id=f.id,
                                                timestamp__gt=last_sent_time,
                                                ).order_by('-timestamp')
-                if stream: # if there are updates to be sent,
-                    last_sent.save() # update timestamp of last sent
-            except LastSent.DoesNotExist: # never updated about this actor, send some updates
-                stream = Action.objects.filter(actor_content_type = content_type,
-                                               actor_object_id = f.id,
-                                               timestamp__gt=datetime.datetime.now()-datetime.timedelta(self.days_back),
+                if stream:  # if there are updates to be sent,
+                    last_sent.save()  # update timestamp of last sent
+            except LastSent.DoesNotExist:  # never updated about this actor, send some updates
+                stream = Action.objects.filter(actor_content_type=content_type,
+                                               actor_object_id=f.id,
+                                               timestamp__gt=datetime.datetime.now() - datetime.timedelta(
+                                                   self.days_back),
                                                ).order_by('-timestamp')
-                last_sent = LastSent.objects.create(user=user,content_type=content_type, object_pk=f.id)
-            if stream: # this actor has some updates
-                try: # genereate the appropriate header for this actor class
-                    header = render_to_string(('notify/%(model)s_header.txt' % {'model': model_template}),{'model':model_name,'object':f})
+                last_sent = LastSent.objects.create(user=user, content_type=content_type, object_pk=f.id)
+            if stream:  # this actor has some updates
+                try:  # genereate the appropriate header for this actor class
+                    header = render_to_string(('notify/%(model)s_header.txt' % {'model': model_template}),
+                                              {'model': model_name, 'object': f})
                 except TemplateDoesNotExist:
-                    header = render_to_string(('notify/model_header.txt'),{'model':model_name,'object':f})
+                    header = render_to_string(('notify/model_header.txt'), {'model': model_name, 'object': f})
                 try:
-                    header_html = render_to_string(('notify/%(model)s_header.html' % {'model': model_template}),{'model':model_name,'object':f,'domain':self.domain})
+                    header_html = render_to_string(('notify/%(model)s_header.html' % {'model': model_template}),
+                                                   {'model': model_name, 'object': f, 'domain': self.domain})
                 except TemplateDoesNotExist:
-                    header_html = render_to_string(('notify/model_header.html'),{'model':model_name,'object':f,'domain':self.domain})
+                    header_html = render_to_string(('notify/model_header.html'),
+                                                   {'model': model_name, 'object': f, 'domain': self.domain})
                 updates[key].append(header)
                 updates_html[key].append(header_html)
 
-                for action_instance in stream: # now generate the updates themselves
+                for action_instance in stream:  # now generate the updates themselves
                     try:
-                        action_output = render_to_string(('activity/%(verb)s/action_email.txt' % { 'verb':action_instance.verb.replace(' ','_') }),{ 'action':action_instance },None)
-                    except TemplateDoesNotExist: # fallback to the generic template
-                        action_output = render_to_string(('activity/action_email.txt'),{ 'action':action_instance },None)
+                        action_output = render_to_string(
+                            ('activity/%(verb)s/action_email.txt' % {'verb': action_instance.verb.replace(' ', '_')}),
+                            {'action': action_instance}, None)
+                    except TemplateDoesNotExist:  # fallback to the generic template
+                        action_output = render_to_string(('activity/action_email.txt'), {'action': action_instance},
+                                                         None)
                     try:
-                        action_output_html = render_to_string(('activity/%(verb)s/action_email.html' % { 'verb':action_instance.verb.replace(' ','_') }),{ 'action':action_instance,'domain':self.domain },None)
-                    except TemplateDoesNotExist: # fallback to the generic template
-                        action_output_html = render_to_string(('activity/action_email.html'),{ 'action':action_instance,'domain':self.domain },None)
+                        action_output_html = render_to_string(
+                            ('activity/%(verb)s/action_email.html' % {'verb': action_instance.verb.replace(' ', '_')}),
+                            {'action': action_instance, 'domain': self.domain}, None)
+                    except TemplateDoesNotExist:  # fallback to the generic template
+                        action_output_html = render_to_string(('activity/action_email.html'),
+                                                              {'action': action_instance, 'domain': self.domain}, None)
                         updates[key].append(action_output)
                     updates_html[key].append(action_output_html)
 
                 if model_class == Agenda:
-                    txt,html = self.agenda_update(f)
+                    txt, html = self.agenda_update(f)
                     updates[key].append(txt)
                     updates_html[key].append(html)
 
         email_body = []
         email_body_html = []
 
-
         # Add the updates for followed models
-        for (model_class,title,title_html) in map(self.get_model_headers, self.update_models):
-            if updates[model_class]: # this model has some updates, add it to the email
+        for (model_class, title, title_html) in map(self.get_model_headers, self.update_models):
+            if updates[model_class]:  # this model has some updates, add it to the email
                 email_body.append(title.format())
                 email_body.append('\n'.join(updates[model_class]))
                 email_body_html.append(title_html.format())
@@ -160,29 +170,29 @@ class Command(NoArgsCommand):
                 party = up.party
                 if party:
                     num_members = cache.get('party_num_members_%d' % party.id,
-                                              None)
+                                            None)
                     if not num_members:
                         num_members = party.userprofile_set.count()
                         cache.set('party_num_members_%d' % party.id,
-                              num_members,
-                              settings.LONG_CACHE_TIME)
+                                  num_members,
+                                  settings.LONG_CACHE_TIME)
                 else:
                     num_members = None
                 debated_bills = get_debated_bills() or []
 
                 template_name = 'notify/party_membership'
                 party_membership_txt = render_to_string(template_name + '.txt',
-                                                        {'user':user,
-                                                         'userprofile':up,
-                                                         'num_members':num_members,
-                                                         'bills':debated_bills,
-                                                         'domain':self.domain})
+                                                        {'user': user,
+                                                         'userprofile': up,
+                                                         'num_members': num_members,
+                                                         'bills': debated_bills,
+                                                         'domain': self.domain})
                 party_membership_html = render_to_string(template_name + '.html',
-                                                         {'user':user,
-                                                         'userprofile':up,
-                                                         'num_members':num_members,
-                                                         'bills':debated_bills,
-                                                         'domain':self.domain})
+                                                         {'user': user,
+                                                          'userprofile': up,
+                                                          'num_members': num_members,
+                                                          'bills': debated_bills,
+                                                          'domain': self.domain})
             else:
                 logger.warning('Can\'t find user profile')
         if email_body:
@@ -191,7 +201,6 @@ class Command(NoArgsCommand):
             email_body_html.insert(0, party_membership_html)
 
         return (email_body, email_body_html)
-
 
     def handle_noargs(self, **options):
 
@@ -212,8 +221,8 @@ class Command(NoArgsCommand):
         queued = 0
         g = Group.objects.get(name='Valid Email')
         for user in User.objects.filter(groups=g,
-                                        profiles__isnull=False)\
-                                .exclude(email=''):
+                                        profiles__isnull=False) \
+                .exclude(email=''):
             try:
                 user_profile = user.profiles.get()
             except UserProfile.DoesNotExist:
@@ -223,16 +232,16 @@ class Command(NoArgsCommand):
                 # if this user has requested emails in the frequency we are
                 # handling now
                 email_body, email_body_html = self.get_email_for_user(user)
-                if email_body: # there are some updates. generate email
-                    header = render_to_string(('notify/header.txt'),{ 'user':user })
-                    footer = render_to_string(('notify/footer.txt'),{ 'user':user,'domain':self.domain })
-                    header_html = render_to_string(('notify/header.html'),{ 'user':user })
-                    footer_html = render_to_string(('notify/footer.html'),{ 'user':user,'domain':self.domain })
+                if email_body:  # there are some updates. generate email
+                    header = render_to_string(('notify/header.txt'), {'user': user})
+                    footer = render_to_string(('notify/footer.txt'), {'user': user, 'domain': self.domain})
+                    header_html = render_to_string(('notify/header.html'), {'user': user})
+                    footer_html = render_to_string(('notify/footer.html'), {'user': user, 'domain': self.domain})
                     send_html_mail(_('Open Knesset Updates'), "%s\n%s\n%s" % (header, '\n'.join(email_body), footer),
-                                                              "%s\n%s\n%s" % (header_html, ''.join(email_body_html), footer_html),
-                                                              self.from_email,
-                                                              [user.email],
-                                                              )
+                                   "%s\n%s\n%s" % (header_html, ''.join(email_body_html), footer_html),
+                                   self.from_email,
+                                   [user.email],
+                                   )
                     queued += 1
 
         logger.info("%d email notifications queued for sending" % queued)
